Time control redesign: Render-Method has now time parameters and as DefaultTime method (gets the time defined by the view) TimeControl class for use in View3D and View2D Remove defective NetVSyncBarrier
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/core/trunk@322 77d55b08-817d-0410-979b-1fec1beb8c2e

+++ view/AbstractCallRender.h
@@ -41,7 +41,7 @@ namespace view {
          *
          * @return The instance time code
          */
-        inline float InstanceTime(void) const {
+        inline double InstanceTime(void) const {
             return this->instTime;
         }
 
@@ -61,7 +61,7 @@ namespace view {
          *
          * @param time The time code of the frame to render
          */
-        inline void SetInstanceTime(float time) {
+        inline void SetInstanceTime(double time) {
             this->instTime = time;
         }
 
@@ -138,7 +138,7 @@ namespace view {
         float time;
 
         /** The instance time code */
-        float instTime;
+        double instTime;
 
         /**
          * Flag marking that 'cntTimeFrames' store the number of the currently
